backend/requirements.txt
Flask>=3.0.0,<4.0.0
flask-cors>=4.0.0,<5.0.0
openai>=1.88.0,<2.0.0
python-dotenv>=1.1.0,<2.0.0